Working on splitting settings duck
