+++ src/App.js
@@ -3,7 +3,8 @@ import React, { useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import ErrorNotifications from "components/ErrorNotifications";
-import { selectIsSettingsLoaded, fetchSettings } from "redux-ducks/settings";
+import { selectIsSettingsLoaded } from "redux-ducks/settings";
+import { fetchSettings } from "redux-ducks/settings/actionCreators";
 import FullScreenLoading from "components/Loading/FullScreenLoading";
 import Router from "routes";
 import "index.css";
